feat(component): 新增公共GroupDescriptons组件
