--- src/pages/deviceMangers/groups/details/pages/groupInfo/index.tsx
@@ -1,5 +1,0 @@
-function GroupInfo() {
-  return <div>GroupInfo</div>;
-}
-
-export default GroupInfo;
